+++ nautilus_trader/common/actor.pyx
@@ -25,13 +25,15 @@ attempts to operate without a managing `Trader` instance.
 """
 
 import warnings
-from typing import Optional
+from typing import Dict, Optional, Set
 
 import cython
 
 from nautilus_trader.config import ActorConfig
+from nautilus_trader.persistence.streaming import generate_signal_class
 
 from cpython.datetime cimport datetime
+from libc.stdint cimport int64_t
 
 from nautilus_trader.cache.base cimport CacheFacade
 from nautilus_trader.common.clock cimport Clock
@@ -67,8 +69,6 @@ from nautilus_trader.model.instruments.base cimport Instrument
 from nautilus_trader.model.orderbook.data cimport OrderBookData
 from nautilus_trader.model.orderbook.data cimport OrderBookSnapshot
 from nautilus_trader.msgbus.bus cimport MessageBus
-
-from nautilus_trader.persistence.streaming import generate_signal_class
 
 
 cdef class Actor(Component):
@@ -108,8 +108,8 @@ cdef class Actor(Component):
             config=config.dict(),
         )
 
-        self._warning_events = set()
-        self._signal_classes = dict()
+        self._warning_events: Set[type] = set()
+        self._signal_classes: Dict[str, type] = {}
 
         self.trader_id = None  # Initialized when registered
         self.msgbus = None     # Initialized when registered
@@ -1320,7 +1320,7 @@ cdef class Actor(Component):
 
         self._msgbus.publish_c(topic=f"data.{data_type.topic}", msg=data)
 
-    cpdef void publish_signal(self, str name, int ts_init, object value, bint stream = False) except *:
+    cpdef void publish_signal(self, str name, value, int64_t ts_event = 0, bint stream = False) except *:
         """
         Publish the given value as a signal to the message bus. Optionally setup persistence for this `signal`.
 
@@ -1329,7 +1329,12 @@ cdef class Actor(Component):
         name : str
             The name of the signal being published.
         value : object
-            The data to publish.
+            The signal data to publish.
+        ts_event : int64, optional
+            The UNIX timestamp (nanoseconds) when the signal event occurred.
+            If ``None`` then will timestamp current time.
+        stream : bool, default False
+            If the signal should also be streamed for persistence.
 
         """
         Condition.not_none(name, "name")
@@ -1337,10 +1342,17 @@ cdef class Actor(Component):
         Condition.is_in(type(value), (int, float, str), "value", "int, float, str")
         Condition.true(self.trader_id is not None, "The actor has not been registered")
 
-        if name not in self._signal_classes:
-            self._signal_classes[name] = generate_signal_class(name=name)
-        cls = self._signal_classes[name]
-        data = cls(ts_init=ts_init, value=value)
+        cdef type cls = self._signal_classes.get(name)
+        if cls is None:
+            cls = generate_signal_class(name=name)
+            self._signal_classes[name] = cls
+
+        cdef int64_t now = self.clock.timestamp_ns()
+        cdef Data data = cls(
+            value=value,
+            ts_event=ts_event or now,
+            ts_init=now,
+        )
         self.publish_data(data_type=DataType(cls), data=data)
 
 # -- REQUESTS -------------------------------------------------------------------------------------